+++ internal/filtering/rewrites_test.go
@@ -57,59 +57,75 @@ func TestRewrites(t *testing.T) {
 	}, {
 		Domain: "a.host3.com",
 		Answer: "x.host.com",
+	}, {
+		Domain: "*.hostboth.com",
+		Answer: "1.2.3.6",
+	}, {
+		Domain: "*.hostboth.com",
+		Answer: "1234::5678",
 	}}
 	d.prepareRewrites()
 
 	testCases := []struct {
 		name      string
 		host      string
-		dtyp      uint16
 		wantCName string
 		wantVals  []net.IP
+		dtyp      uint16
 	}{{
-		name: "not_filtered_not_found",
-		host: "hoost.com",
-		dtyp: dns.TypeA,
+		name:      "not_filtered_not_found",
+		host:      "hoost.com",
+		wantCName: "",
+		wantVals:  nil,
+		dtyp:      dns.TypeA,
 	}, {
 		name:      "rewritten_a",
 		host:      "www.host.com",
-		dtyp:      dns.TypeA,
 		wantCName: "host.com",
 		wantVals:  []net.IP{{1, 2, 3, 4}, {1, 2, 3, 5}},
+		dtyp:      dns.TypeA,
 	}, {
 		name:      "rewritten_aaaa",
 		host:      "www.host.com",
-		dtyp:      dns.TypeAAAA,
 		wantCName: "host.com",
 		wantVals:  []net.IP{net.ParseIP("1:2:3::4")},
+		dtyp:      dns.TypeAAAA,
 	}, {
-		name:     "wildcard_match",
-		host:     "abc.host.com",
-		dtyp:     dns.TypeA,
-		wantVals: []net.IP{{1, 2, 3, 5}},
+		name:      "wildcard_match",
+		host:      "abc.host.com",
+		wantCName: "",
+		wantVals:  []net.IP{{1, 2, 3, 5}},
+		dtyp:      dns.TypeA,
 	}, {
-		name:     "wildcard_override",
-		host:     "a.host.com",
-		dtyp:     dns.TypeA,
-		wantVals: []net.IP{{1, 2, 3, 4}},
+		name:      "wildcard_override",
+		host:      "a.host.com",
+		wantCName: "",
+		wantVals:  []net.IP{{1, 2, 3, 4}},
+		dtyp:      dns.TypeA,
 	}, {
 		name:      "wildcard_cname_interaction",
 		host:      "www.host2.com",
-		dtyp:      dns.TypeA,
 		wantCName: "host.com",
 		wantVals:  []net.IP{{1, 2, 3, 4}, {1, 2, 3, 5}},
+		dtyp:      dns.TypeA,
 	}, {
 		name:      "two_cnames",
 		host:      "b.host.com",
-		dtyp:      dns.TypeA,
 		wantCName: "somehost.com",
 		wantVals:  []net.IP{{0, 0, 0, 0}},
+		dtyp:      dns.TypeA,
 	}, {
 		name:      "two_cnames_and_wildcard",
 		host:      "b.host3.com",
-		dtyp:      dns.TypeA,
 		wantCName: "x.host.com",
 		wantVals:  []net.IP{{1, 2, 3, 5}},
+		dtyp:      dns.TypeA,
+	}, {
+		name:      "issue3343",
+		host:      "www.hostboth.com",
+		wantCName: "",
+		wantVals:  []net.IP{net.ParseIP("1234::5678")},
+		dtyp:      dns.TypeAAAA,
 	}}
 
 	for _, tc := range testCases {
@@ -143,12 +159,15 @@ func TestRewritesLevels(t *testing.T) {
 	d.Rewrites = []RewriteEntry{{
 		Domain: "host.com",
 		Answer: "1.1.1.1",
+		Type:   dns.TypeA,
 	}, {
 		Domain: "*.host.com",
 		Answer: "2.2.2.2",
+		Type:   dns.TypeA,
 	}, {
 		Domain: "*.sub.host.com",
 		Answer: "3.3.3.3",
+		Type:   dns.TypeA,
 	}}
 	d.prepareRewrites()
 
@@ -234,53 +253,61 @@ func TestRewritesExceptionIP(t *testing.T) {
 	d.Rewrites = []RewriteEntry{{
 		Domain: "host.com",
 		Answer: "1.2.3.4",
+		Type:   dns.TypeA,
 	}, {
 		Domain: "host.com",
 		Answer: "AAAA",
+		Type:   dns.TypeAAAA,
 	}, {
 		Domain: "host2.com",
 		Answer: "::1",
+		Type:   dns.TypeAAAA,
 	}, {
 		Domain: "host2.com",
 		Answer: "A",
+		Type:   dns.TypeA,
 	}, {
 		Domain: "host3.com",
 		Answer: "A",
+		Type:   dns.TypeA,
 	}}
 	d.prepareRewrites()
 
 	testCases := []struct {
 		name string
 		host string
-		dtyp uint16
 		want []net.IP
+		dtyp uint16
 	}{{
 		name: "match_A",
 		host: "host.com",
-		dtyp: dns.TypeA,
 		want: []net.IP{{1, 2, 3, 4}},
+		dtyp: dns.TypeA,
 	}, {
 		name: "exception_AAAA_host.com",
 		host: "host.com",
+		want: nil,
 		dtyp: dns.TypeAAAA,
 	}, {
 		name: "exception_A_host2.com",
 		host: "host2.com",
+		want: nil,
 		dtyp: dns.TypeA,
 	}, {
 		name: "match_AAAA_host2.com",
 		host: "host2.com",
-		dtyp: dns.TypeAAAA,
 		want: []net.IP{net.ParseIP("::1")},
+		dtyp: dns.TypeAAAA,
 	}, {
 		name: "exception_A_host3.com",
 		host: "host3.com",
+		want: nil,
 		dtyp: dns.TypeA,
 	}, {
 		name: "match_AAAA_host3.com",
 		host: "host3.com",
+		want: nil,
 		dtyp: dns.TypeAAAA,
-		want: []net.IP{},
 	}}
 
 	for _, tc := range testCases {
@@ -293,7 +320,9 @@ func TestRewritesExceptionIP(t *testing.T) {
 			}
 
 			assert.Equal(t, Rewritten, r.Reason)
+
 			require.Len(t, r.IPList, len(tc.want))
+
 			for _, ip := range tc.want {
 				assert.True(t, ip.Equal(r.IPList[0]))
 			}